renaming the utils and fixing import in max_current_monitor.py
